[Upstream]gsupplicant: Switch to different BSSID of same SSID in case of signal bss removed
Teach ConnMan to select the next BSSID when two APs (e.g. 2.4 GHz and
5 GHz) with the same SSID and security configuration are available.

Currently when one AP disappears ConnMan will retry to connect to the
old AP point even though there is another matching BSSID available.

So when wpa_supplicant sends a remove signal we find a matching BSSID
and use it even when it's not from the same AP.
https://git.kernel.org/pub/scm/network/connman/connman.git/commit/?id=c1debcac758ef1f76a788225679003e2e67d9fdf

Change-Id: Ic5ce438addc823ebd0c6bedd2a974f559b33ae25
Signed-off-by: Niraj Kumar Goit <niraj.g@samsung.com>